Clarify ADDRESS_ASSIGN, ROUTE_ADVERTISEMENT roles
This change adds text that addresses conveyed by ADDRESS_ASSIGN can be
used in the "source address" field of an IP packet, while addresses
conveyed by ROUTE_ADVERTISEMENT can be used in the "destination
address" field.

While we're here, also add a small clarification when a prefix is
assigned/advertised that *any* of these addresses can be used in the
source/destination field, respectively. A future extension could
create semantics for these addreses, but that is not called out in the
text at this time.

Closes: #25